pyproject.toml
[project]
name = "cognitive-book-os"
version = "0.1.0"
description = "A structured knowledge extraction system for documents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.75.0",
    "fastapi>=0.128.0",
    "instructor>=1.13.0",
    "langfuse>=2.0.0",
    "openai>=2.14.0",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "pyyaml>=6.0.0",
    "rich>=14.2.0",
    "streamlit>=1.52.2",
    "typer>=0.21.0",
    "uvicorn>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[project.scripts]
bookos = "cognitive_book_os.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cognitive_book_os"]
